+++ QuizUp.Server/Controllers/QuizzesController.cs
@@ -3,12 +3,13 @@ using Microsoft.AspNetCore.Mvc;
 using QuizUp.BL.Exceptions;
 using QuizUp.BL.Services;
 using QuizUp.BL.Models;
+using System.Security.Claims;
 
 namespace QuizUp.Server.Controllers;
 
 [Route("api/quizzes")]
 [ApiController]
-//[Authorize]
+[Authorize]
 public class QuizzesController(IQuizService quizService) : ControllerBase
 {
     private ActionResult InternalServerError =>
@@ -17,6 +18,12 @@ public class QuizzesController(IQuizService quizService) : ControllerBase
     [HttpGet]
     public async Task<ActionResult<List<QuizSummaryModel>>> GetQuizzesByUserIdAsync([FromQuery] Guid userId)
     {
+        var accessTokenUserId = GetAccessTokenUserId();
+        if (accessTokenUserId == null || accessTokenUserId != userId)
+        {
+            return Unauthorized();
+        }
+
         try
         {
             var quizSummaryModels = await quizService.GetQuizzessByUserIdAsync(userId);
@@ -35,6 +42,12 @@ public class QuizzesController(IQuizService quizService) : ControllerBase
     [HttpGet("{id:Guid}")]
     public async Task<ActionResult<QuizDetailModel>> GetQuizByIdAsync(Guid id)
     {
+        var quizBelongsToUser = await DoesQuizBelongToUser(id);
+        if (!quizBelongsToUser)
+        {
+            return Unauthorized();
+        }
+
         try
         {
             var quizDetailModel = await quizService.GetQuizByIdAsync(id);
@@ -53,6 +66,12 @@ public class QuizzesController(IQuizService quizService) : ControllerBase
     [HttpPost]
     public async Task<ActionResult<QuizDetailModel>> CreateQuizAsync([FromBody] CreateQuizModel createQuizModel)
     {
+        var accessTokenUserId = GetAccessTokenUserId();
+        if (accessTokenUserId == null || accessTokenUserId != createQuizModel.UserId)
+        {
+            return Unauthorized();
+        }
+
         try
         {
             var quizDetailModel = await quizService.CreateQuizAsync(createQuizModel);
@@ -67,6 +86,12 @@ public class QuizzesController(IQuizService quizService) : ControllerBase
     [HttpPut("{id:Guid}")]
     public async Task<ActionResult> EditQuizAsync(Guid id, [FromBody] EditQuizModel editQuizModel)
     {
+        var quizBelongsToUser = await DoesQuizBelongToUser(id);
+        if (!quizBelongsToUser)
+        {
+            return Unauthorized();
+        }
+
         try
         {
             await quizService.EditQuizAsync(id, editQuizModel);
@@ -85,6 +110,12 @@ public class QuizzesController(IQuizService quizService) : ControllerBase
     [HttpDelete("{id:Guid}")]
     public async Task<ActionResult> DeleteQuizAsync(Guid id)
     {
+        var quizBelongsToUser = await DoesQuizBelongToUser(id);
+        if (!quizBelongsToUser)
+        {
+            return Unauthorized();
+        }
+
         try
         {
             await quizService.DeleteQuizByIdAsync(id);
@@ -104,6 +135,12 @@ public class QuizzesController(IQuizService quizService) : ControllerBase
     [HttpGet("games/{id:Guid}")]
     public async Task<ActionResult<QuizGamesModel>> GetGamesByQuizIdAsync(Guid id)
     {
+        var quizBelongsToUser = await DoesQuizBelongToUser(id);
+        if (!quizBelongsToUser)
+        {
+            return Unauthorized();
+        }
+
         try
         {
             var quizGamesModel = await quizService.GetGamesByQuizIdAsync(id);
@@ -117,5 +154,27 @@ public class QuizzesController(IQuizService quizService) : ControllerBase
         {
             return InternalServerError;
         }
+    }
+
+    private Guid? GetAccessTokenUserId()
+    {
+        var userIdString = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if (userIdString == null)
+        {
+            return null;
+        }
+
+        return new Guid(userIdString);
+    }
+
+    private async Task<bool> DoesQuizBelongToUser(Guid quizId)
+    {
+        var userId = GetAccessTokenUserId();
+        if (userId == null)
+        {
+            return false;
+        }
+
+        return await quizService.DoesQuizBelongToUser(quizId, userId ?? Guid.Empty);
     }
 }
